+++ tests/kernel/wave/nn/functional/wave_sdpa_test.py
@@ -12,8 +12,8 @@ import iree.turbine.kernel.wave.nn as wave_nn
 from ...common.utils import (
     require_e2e,
     require_cdna3,
-    scaled_dot_product_attention_bhsd,
 )
+from iree.turbine.kernel.wave.common.utils import scaled_dot_product_attention_bhsd
 
 
 @require_e2e
